+++ lib/categories/view/categories_page.dart
@@ -1,3 +1,4 @@
+import 'package:awesome_snackbar_content/awesome_snackbar_content.dart';
 import 'package:qruto_budget/categories/cubit/categories_cubit.dart';
 import 'package:qruto_budget/categories/repository/category_repository.dart';
 import 'package:flutter/material.dart';
@@ -7,6 +8,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
 import 'package:qruto_budget/database/database.dart';
 
+import '../../shared/shared.dart';
 import '../../transaction/models/transaction_type.dart';
 import '../../utils/theme/budget_theme.dart';
 import '../../utils/theme/cubit/theme_cubit.dart';
@@ -37,13 +39,8 @@ class CategoriesView extends StatelessWidget {
       listenWhen: (previous, current) => previous.status != current.status,
       listener: (context, state) {
         if (state.status == CategoriesStatus.failure) {
-          ScaffoldMessenger.of(context)
-            ..hideCurrentSnackBar()
-            ..showSnackBar(
-              SnackBar(
-                content: Text(state.errorMessage!),
-              ),
-            );
+          SharedFunctions.showSnackbar(
+              context, false, 'Ups', state.errorMessage!);
         }
       },
       builder: (context, state) {