+++ lib/shared/shared.dart
@@ -1,1 +1,4 @@
 export 'widgets/widgets.dart';
+export 'shared_snackbar.dart';
+export 'background_worker.dart';
+export 'awesome_icons.dart';