+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - Select a custom range in the Summary
 - Predefined income or expense categories
+### Fixed
+- Category or Account can't be deleted if they contain any transaction
 
 ## [4.4.0] - 03-26-2025
 ### Added